requirements.txt
# Conexão ao banco de dados
sshtunnel==0.4.0
pymongo==4.11

# Credenciais
PyYAML==6.0.2

# Vetorização e/ou classificação das notícias
scikit-learn==1.6.1
gensim==4.3.3
imbalanced-learn==0.13.0

# Exibição
tqdm==4.67.1

# LLM e Processamento de Linguagem Natural
google-generativeai>=0.4.1
pydantic>=2.0.0